Move RuntimeOptionParser to guest staging
