+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/DumpIsolateCreationOnlyOptionsFeature.java
@@ -38,7 +38,7 @@ import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.shared.option.HostedOptionKey;
 import com.oracle.svm.guest.staging.option.RuntimeOptionKey;
-import com.oracle.svm.core.option.RuntimeOptionParser;
+import com.oracle.svm.guest.staging.option.RuntimeOptionParser;
 import com.oracle.svm.shared.util.VMError;
 
 import jdk.graal.compiler.options.Option;
